Replace Sharding references with Tenant
