+++ src/backend/app/modules/Tenant/ApiResolvers/AfricasTalkingApiResolver.php
@@ -1,18 +1,20 @@
 <?php
 
-namespace MPM\Sharding\ApiResolvers;
+declare(strict_types=1);
+
+namespace MPM\Tenant\ApiResolvers;
 
 use App\Exceptions\ValidationException;
 use Illuminate\Http\Request;
 
-class AndroidGatewayCallbackApiResolver implements ApiResolverInterface {
+class AfricasTalkingApiResolver implements ApiResolverInterface {
     public function resolveCompanyId(Request $request): int {
         $segments = $request->segments();
         if (count($segments) !== 5) {
             throw new ValidationException('failed to parse company identifier from the webhook');
         }
 
-        $companyId = $segments[4];
+        $companyId = $segments[3];
 
         return (int) $companyId;
     }
