+++ src/backend/app/modules/Tenant/ApiResolvers/Data/ApiResolverMap.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace MPM\Sharding\ApiResolvers\Data;
+namespace MPM\Tenant\ApiResolvers\Data;
 
-use MPM\Sharding\ApiResolvers\AfricasTalkingApiResolver;
-use MPM\Sharding\ApiResolvers\AndroidGatewayCallbackApiResolver;
-use MPM\Sharding\ApiResolvers\DataExportResolver;
-use MPM\Sharding\ApiResolvers\DownloadingReportsResolver;
-use MPM\Sharding\ApiResolvers\SwiftaPaymentApiResolver;
-use MPM\Sharding\ApiResolvers\TestApiResolver;
-use MPM\Sharding\ApiResolvers\ViberMessagingApiResolver;
-use MPM\Sharding\ApiResolvers\WaveMoneyApiResolver;
+use MPM\Tenant\ApiResolvers\AfricasTalkingApiResolver;
+use MPM\Tenant\ApiResolvers\AndroidGatewayCallbackApiResolver;
+use MPM\Tenant\ApiResolvers\DataExportResolver;
+use MPM\Tenant\ApiResolvers\DownloadingReportsResolver;
+use MPM\Tenant\ApiResolvers\SwiftaPaymentApiResolver;
+use MPM\Tenant\ApiResolvers\TestApiResolver;
+use MPM\Tenant\ApiResolvers\ViberMessagingApiResolver;
+use MPM\Tenant\ApiResolvers\WaveMoneyApiResolver;
 
 class ApiResolverMap {
     public const VIBER_API = 'api/viber-messaging/webhook';
