+++ src/backend/app/modules/Tenant/ApiResolvers/TestApiResolver.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace MPM\Sharding\ApiResolvers;
+namespace MPM\Tenant\ApiResolvers;
 
 use App\Exceptions\ValidationException;
 use Illuminate\Http\Request;
